See below
Inverse has been split into "invert" and "inverse", invert flips the
stack in reference, inverse produces a new flipped stack from the
reference. Both are optimised

Split implimented, which returns a stack consisting of parts that match
the top value's pattern. tostack implements this with the pattern '.'

Lua Implementation has been changed to a sub folder, instead of a zip.

Added default code to the gitignore