Ease Changes
while and if no-longer hold onto their conditional, unless _peek is
added to the end. These can be re-asoc'd at the start of code anyway, if
that's the preferred functionality.

rand now pushes a random value of the stack underneith it, if it's a
stack and not two numbers.

Added 'sum', which either sums up all the values in a stack via the '+'
operator, (Which also works for strings to re-concatenate them), or just
acts like the '+' operator.