Math now works on some other things
Strings can be multiplied, which, unless the string can be cast to a
number, acts as repeat.
Functions can be multiplied, which returns a new function that consists
of the old function running 'n' times. Not sure a good use of this just
yet.
Functions can be added together, which chains functions if you're into
that.