Bunch of Quality of Life stuff
Added `#`, which when at the start of a word, or before a character in a
ZSS, returns the function the rest of the word represents. `##word` will
produce a function which pushes `word`. Etc.

Loads of sugar added, This has been needed for some time. Now we're down
to a tiny collection of 1 byte values yet to be used.

Has value now falls back to `value stack` if `value` isn't a stack.

Arithmatic now works across two stacks. Try `( 1 2 3 ) ( 4 5 6 ) +`

`base` and `frombase` now support negative bases for the inverse action.